split basic and advance components
