+++ apps/fishing-map/features/search/SearchResults.module.css
@@ -1,0 +1,75 @@
+.searchResult {
+  position: relative;
+  padding: var(--space-M) var(--space-S);
+  background-color: transparent;
+  cursor: pointer;
+  border-left: 0.1rem solid transparent;
+  border-right: 0.5rem solid transparent;
+  transition: background-color 150ms linear, box-shadow 150ms linear, border 150ms linear;
+}
+
+.searchResult:hover,
+.highlighted {
+  background-color: var(--color-white);
+  box-shadow: var(--box-shadow);
+  border-left: 0.5rem solid var(--color-secondary-blue);
+  border-right: 0.1rem solid transparent;
+}
+
+.selected {
+  background: var(--color-terthiary-blue);
+}
+
+.inWorkspace {
+  /* background: var(--color-terthiary-blue); */
+  color: var(--color-secondary-blue);
+  pointer-events: none;
+}
+
+.alreadyAddedMsg {
+  display: flex;
+  align-items: center;
+  font: var(--font-S);
+}
+
+.searchResult:not(:last-child)::after {
+  content: '';
+  position: absolute;
+  left: var(--space-S);
+  bottom: 0;
+  height: 0.1rem;
+  width: calc(100% - var(--space-S));
+  background-color: var(--color-border);
+  z-index: -1;
+}
+
+.properties {
+  display: flex;
+  flex-wrap: wrap;
+  opacity: var(--opacity-secondary);
+  transition: opacity 150ms linear;
+}
+
+.property {
+  margin-right: var(--space-M);
+  margin-top: var(--space-S);
+}
+
+.property label {
+  line-height: 1;
+}
+
+.property span {
+  white-space: nowrap;
+}
+
+.fullWidth {
+  margin-right: unset;
+  width: 100%;
+}
+
+.searchResult:hover .properties,
+.searchResult:hover .properties label {
+  cursor: pointer;
+  opacity: var(--opacity-primary);
+}
